+++ frameworks/napi/src/js_register_handle.cpp
@@ -15,6 +15,7 @@
 #include "js_register_handle.h"
 #include "js_register_event.h"
 #include "mmi_token.h"
+#include <inttypes.h>
 
 namespace OHOS {
     namespace MMI {
@@ -200,7 +201,7 @@ namespace OHOS {
             std::string registerHandle = std::to_string(winId) + "," + std::to_string(type);
             g_registerMap.insert(std::pair<std::string, RegisterHanldeInfo>(registerHandle, registerInfo));
             HILOG_DEBUG("JSRegisterHandle::Register: registerHandle=%{public}s", registerHandle.c_str());
-            HILOG_DEBUG("JSRegisterHandle::Register: registerMap size=%{public}d", g_registerMap.size());
+            HILOG_DEBUG("JSRegisterHandle::Register: registerMap size=%{public}d", static_cast<int32_t>(g_registerMap.size()));
             HILOG_DEBUG("JSRegisterHandle::Register: success");
             return response;
         }
@@ -221,7 +222,7 @@ namespace OHOS {
                 HILOG_DEBUG("JSRegisterHandle::Unregister: registerHandle=%{public}s", registerHandle.c_str());
                 g_registerMap.erase(iter);
             }
-            HILOG_DEBUG("JSRegisterHandle::Unregister: registerMap size=%{public}d", g_registerMap.size());
+            HILOG_DEBUG("JSRegisterHandle::Unregister: registerMap size=%{public}d", static_cast<int32_t>(g_registerMap.size()));
             HILOG_DEBUG("JSRegisterHandle::Unregister: success");
             return response;
         }
@@ -241,7 +242,7 @@ namespace OHOS {
                 HILOG_DEBUG("JSRegisterHandle::UnregisterAll: registerHandle=%{public}s", iter->first.c_str());
                 g_registerMap.erase(iter);
             }
-            HILOG_DEBUG("JSRegisterHandle::UnregisterAll: registerMap size=%{public}d", g_registerMap.size());
+            HILOG_DEBUG("JSRegisterHandle::UnregisterAll: registerMap size=%{public}d", static_cast<int32_t>(g_registerMap.size()));
             HILOG_DEBUG("JSRegisterHandle::UnregisterAll: success");
             return response;
         }
